GenerativeAI/chapter4/pyproject.toml
[project]
name = "chapter4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.15",
    "langchain-community==0.4.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "python-dotenv==1.2.1",
]
